Fixed new warnings from GCC

+++ lib/my/src/stdio/internal/file/do_read.c
@@ -13,7 +13,7 @@
 static void do_offset(my_file_t *fp, ssize_t read_result)
 {
     if (fp->flags & MY_FILE_FLAG_IS_OFFSET_CORRECT) {
-        if (fp->offset <= MY_NUMERIC_LIMITS_MAX(off_t) - read_result)
+        if ((size_t)fp->offset <= MY_NUMERIC_LIMITS_MAX(off_t) - read_result)
             fp->offset += read_result;
         else
             fp->flags &= ~MY_FILE_FLAG_IS_OFFSET_CORRECT;
